[GTK] Scrollbars not correctly rendered in non GNOME environments https://bugs.webkit.org/show_bug.cgi?id=153860
Reviewed by Michael Catanzaro.

I noticed this in a matchbox environment, where there's no
gnome-setting-daemon running. The problem is only with the
scrollbars, because we initialize the GtkSettings in
RenderThemeGtk and notify the ScrollbarTheme when it changes, but
ScrollbarTheme is created before RenderThemeGtk so we initialize
the theme properties before the GtkSettings have been
initialized. We can just let the ScrollbarTheme monitor the
theme itself instead of relying on being notified by the WebCore
layer.

* platform/gtk/ScrollbarThemeGtk.cpp:
(WebCore::themeChangedCallback):
(WebCore::ScrollbarThemeGtk::ScrollbarThemeGtk):
* rendering/RenderThemeGtk.cpp:
(WebCore::gtkStyleChangedCallback): Deleted.

Identifier: 172009@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@196171 268f45cc-cd09-0410-ab3c-d52691b4dbfc